backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
# OLM OCR Libraries - Compatible with Python 3.9
transformers==4.35.2
torch==2.1.1+cu121
torchvision==0.16.1+cu121
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
accelerate==0.24.1
sentencepiece==0.1.99
# Avoid TensorFlow conflicts by pinning protobuf
protobuf==3.20.3
# Force specific versions to avoid conflicts
typing-extensions==4.8.0 